removed second src folder
