--- src/components/ContactForm/FormContainer/index.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Form from "../Form";
-
-function FormContainer() {
-  return(
-    <div>
-      <h2>Contact SkyLine</h2>
-      <Form />
-    </div>
-  )
-}
-
-export default FormContainer;
