--- src/components/ContactForm/SubmitButton/index.js
@@ -1,9 +1,0 @@
-import React from "react"
-
-function SubmitButton() {
-  return(
-    <button type="submit" id="contact-submit-btn">Send</button>
-  )
-}
-
-export default SubmitButton;
